+++ CHANGELOG.md
@@ -18,9 +18,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   dependencies (i.e. when the `bindeps` feature is used) by default
 * `inheritCargoArtifactsHook` will now symlink dependency `.rlib` and `.rmeta`
   files. This means that derivations which reuse existing cargo artifacts will
-  run faster as fewer files (and bytes!) need to be copied around
-* `cargoTarpaulin`'s default `cargoTarpaulinExtraArgs` no longer include
-  `--skip-clean`
+  run faster as fewer files (and bytes!) need to be copied around. To disable
+  this behavior, set `doNotLinkInheritedArtifacts = true;`.
+* `cargoTarpaulin` will now set `doNotLinkInheritedArtifacts = true;` unless
+  otherwise specified
 
 ### Changed
 * **Breaking**: dropped compatibility for Nix versions below 2.13.3