+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   this behavior, set `doNotLinkInheritedArtifacts = true;`.
 * `cargoTarpaulin` will now set `doNotLinkInheritedArtifacts = true;` unless
   otherwise specified
+* Update `crane-utils` dependencies for successful build in nightly Rust (2023-06-28)
 
 ### Changed
 * **Breaking**: dropped compatibility for Nix versions below 2.13.3