+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### [0.13.0] - 2023-08-07
+
 ### Added
 * `buildPackage` now supports installing `dylib` targets
 * Added support for sparse registries
@@ -432,6 +434,7 @@ files parsed as nix attribute sets.
 ## 0.1.0 - 2022-01-22
 - First release
 
+[0.13.0]: https://github.com/ipetkov/crane/compare/v0.12.2...v0.13.0
 [0.12.2]: https://github.com/ipetkov/crane/compare/v0.12.1...v0.12.2
 [0.12.1]: https://github.com/ipetkov/crane/compare/v0.12.0...v0.12.1
 [0.12.0]: https://github.com/ipetkov/crane/compare/v0.11.3...v0.12.0